requirements.txt
numpy==1.23.5
scikit-learn==1.2.2
pandas==1.5.3
flask==2.3.2
gunicorn==20.1.0
python-dotenv==1.0.0